feat: Implemented showing my rentals.
refactor: Moved UserProfileSection.kt to a different package.
